pyproject.toml
[project]
name = "jq-screening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "jquants-api-client>=1.8.0",
    "pandas>=2.3.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
main = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
